requirements.txt
flask
langchain
langchain_community
sentence-transformers
chromadb
pypdf
python-dotenv

-e .